Some test restructuring
