+++ automation/src/test/scala/org/broadinstitute/dsde/workbench/leonardo/ClusterMonitoringSpec.scala
@@ -2,7 +2,7 @@ package org.broadinstitute.dsde.workbench.leonardo
 
 import java.io.File
 
-import org.broadinstitute.dsde.workbench.service.{Orchestration, Sam}
+import org.broadinstitute.dsde.workbench.service.{Orchestration, RestException, Sam}
 import org.broadinstitute.dsde.workbench.dao.Google.{googleIamDAO, googleStorageDAO}
 import org.broadinstitute.dsde.workbench.fixture.BillingFixtures
 import org.broadinstitute.dsde.workbench.model.google.GcsEntityTypes.Group
@@ -11,8 +11,11 @@ import org.broadinstitute.dsde.workbench.model.google.{GcsEntity, GcsObjectName,
 import org.scalatest.{FreeSpec, ParallelTestExecution}
 
 class ClusterMonitoringSpec extends FreeSpec with LeonardoTestUtils with ParallelTestExecution with BillingFixtures {
-  "Leonardo clusters" - {
 
+  // these tests just hit the Leo APIs; they don't interact with notebooks via selenium
+  "Cluster status transitions" - {
+
+    // (create -> wait -> delete -> wait) * 2
     "should create, monitor, delete, recreate, and re-delete a cluster" in {
       withCleanBillingProject(hermioneCreds) { projectName =>
         Orchestration.billing.addUserToBillingProject(projectName, ronEmail, Orchestration.billing.BillingProjectRole.User)(hermioneAuthToken)
@@ -37,6 +40,122 @@ class ClusterMonitoringSpec extends FreeSpec with LeonardoTestUtils with Paralle
         }
       }
     }
+
+    // create -> no wait -> delete
+    "should delete a creating cluster" in withWebDriver { implicit driver =>
+      withCleanBillingProject(hermioneCreds) { projectName =>
+        Orchestration.billing.addUserToBillingProject(projectName, ronEmail, Orchestration.billing.BillingProjectRole.User)(hermioneAuthToken)
+        val project = GoogleProject(projectName)
+        implicit val token = ronAuthToken
+
+        withNewCluster(project, monitorCreate = false, monitorDelete = true)(_ => ())
+      }
+    }
+
+    // create -> create again (conflict) -> delete
+    "should not be able to create 2 clusters with the same name" in withWebDriver { implicit driver =>
+      withCleanBillingProject(hermioneCreds) { projectName =>
+        Orchestration.billing.addUserToBillingProject(projectName, ronEmail, Orchestration.billing.BillingProjectRole.User)(hermioneAuthToken)
+        val project = GoogleProject(projectName)
+        implicit val token = ronAuthToken
+
+        withNewCluster(project, monitorCreate = false, monitorDelete = true) { cluster =>
+          val caught = the [RestException] thrownBy createNewCluster(project, cluster.clusterName, monitor = false)
+          caught.message should include (""""statusCode":409""")
+        }
+      }
+    }
+
+    // create -> wait -> delete -> no wait -> create (conflict)
+    // TODO fails due to "statusCode":500.
+    // rogue actor running
+    "should not be able to recreate a deleting cluster" in withWebDriver { implicit driver =>
+      withCleanBillingProject(hermioneCreds) { projectName =>
+        Orchestration.billing.addUserToBillingProject(projectName, ronEmail, Orchestration.billing.BillingProjectRole.User)(hermioneAuthToken)
+        val project = GoogleProject(projectName)
+        implicit val token = ronAuthToken
+
+        val cluster = withNewCluster(project, monitorCreate = true, monitorDelete = false)(identity)
+
+        val caught = the [RestException] thrownBy createNewCluster(project, cluster.clusterName)
+        caught.message should include (""""statusCode":409""")
+      }
+    }
+
+    // create -> wait -> delete -> no wait -> delete (conflict)
+    // TODO failed, no exception thrown
+    // second deletion seemed to have no effect
+    "should not be able to delete a deleting cluster" in withWebDriver { implicit driver =>
+      withCleanBillingProject(hermioneCreds) { projectName =>
+        Orchestration.billing.addUserToBillingProject(projectName, ronEmail, Orchestration.billing.BillingProjectRole.User)(hermioneAuthToken)
+        val project = GoogleProject(projectName)
+        implicit val token = ronAuthToken
+
+        val cluster = withNewCluster(project, monitorCreate = true, monitorDelete = false)(identity)
+
+        val caught = the [RestException] thrownBy deleteCluster(project, cluster.clusterName, monitor = false)
+        caught.message should include (""""statusCode":409""")
+      }
+    }
+
+    // create -> no wait -> stop (conflict) -> delete
+    // TODO failed, no exception thrown
+    "should not be able to stop a creating cluster" in withWebDriver { implicit driver =>
+      withCleanBillingProject(hermioneCreds) { projectName =>
+        Orchestration.billing.addUserToBillingProject(projectName, ronEmail, Orchestration.billing.BillingProjectRole.User)(hermioneAuthToken)
+        val project = GoogleProject(projectName)
+        implicit val token = ronAuthToken
+
+        withNewCluster(project, monitorCreate = false, monitorDelete = true) { cluster =>
+          val caught = the[RestException] thrownBy stopCluster(project, cluster.clusterName, monitor = false)
+          caught.message should include (""""statusCode":409""")
+        }
+      }
+    }
+
+    // create -> wait -> stop -> no wait -> start -> delete
+    "should be able to start a stopping cluster" in withWebDriver { implicit driver =>
+      withCleanBillingProject(hermioneCreds) { projectName =>
+        Orchestration.billing.addUserToBillingProject(projectName, ronEmail, Orchestration.billing.BillingProjectRole.User)(hermioneAuthToken)
+        val project = GoogleProject(projectName)
+        implicit val token = ronAuthToken
+
+        withNewCluster(project, monitorCreate = true, monitorDelete = false) { cluster =>
+          stopCluster(project, cluster.clusterName, monitor = false)
+          startCluster(project, cluster.clusterName, monitor = false)
+        }
+      }
+    }
+
+    // create -> wait -> stop -> wait -> delete
+    "should be able to delete a stopped cluster" in withWebDriver { implicit driver =>
+      withCleanBillingProject(hermioneCreds) { projectName =>
+        Orchestration.billing.addUserToBillingProject(projectName, ronEmail, Orchestration.billing.BillingProjectRole.User)(hermioneAuthToken)
+        val project = GoogleProject(projectName)
+        implicit val token = ronAuthToken
+
+        withNewCluster(project, monitorCreate = true, monitorDelete = true) { cluster =>
+          stopAndMonitor(cluster.googleProject, cluster.clusterName)
+        }
+      }
+    }
+
+    // create -> wait -> stop -> no wait -> delete
+    "should be able to delete a stopping cluster" in withWebDriver { implicit driver =>
+      withCleanBillingProject(hermioneCreds) { projectName =>
+        Orchestration.billing.addUserToBillingProject(projectName, ronEmail, Orchestration.billing.BillingProjectRole.User)(hermioneAuthToken)
+        val project = GoogleProject(projectName)
+        implicit val token = ronAuthToken
+
+        withNewCluster(project, monitorCreate = true, monitorDelete = true) { cluster =>
+          stopCluster(cluster.googleProject, cluster.clusterName, monitor = false)
+        }
+      }
+    }
+  }
+
+  // these tests interact with the Leo APIs _and_ interact with notebooks via selenium
+  "Leonardo clusters" - {
 
     // default PetClusterServiceAccountProvider edition
     "should create a cluster in a different billing project using PetClusterServiceAccountProvider and put the pet's credentials on the cluster" in withWebDriver { implicit driver =>
@@ -177,23 +296,6 @@ class ClusterMonitoringSpec extends FreeSpec with LeonardoTestUtils with Paralle
       }
     }
 
-    "should be able to delete a stopped cluster" in withWebDriver { implicit driver =>
-      withCleanBillingProject(hermioneCreds) { projectName =>
-        val project = GoogleProject(projectName)
-
-        Orchestration.billing.addUserToBillingProject(projectName, ronEmail, Orchestration.billing.BillingProjectRole.User)(hermioneAuthToken)
-
-        implicit val token = ronAuthToken
-
-        // Create a cluster
-        withNewCluster(project) { cluster =>
-          // Stop the cluster
-          stopAndMonitor(cluster.googleProject, cluster.clusterName)
-        }
-        // Delete should succeed
-      }
-    }
-
     "should pause and resume a cluster with preemptible instances" in withWebDriver { implicit driver =>
       withCleanBillingProject(hermioneCreds) { projectName =>
         val project = GoogleProject(projectName)
@@ -240,6 +342,26 @@ class ClusterMonitoringSpec extends FreeSpec with LeonardoTestUtils with Paralle
         }
       }
 
+    }
+
+    //Test to check if extensions are installed correctly
+    //Using nbtranslate extension from here:
+    //https://github.com/ipython-contrib/jupyter_contrib_nbextensions/tree/master/src/jupyter_contrib_nbextensions/nbextensions/nbTranslate
+    "should install user specified notebook extensions" in withWebDriver { implicit driver =>
+      withCleanBillingProject(hermioneCreds) { projectName =>
+        Orchestration.billing.addUserToBillingProject(projectName, ronEmail, Orchestration.billing.BillingProjectRole.User)(hermioneAuthToken)
+        val project = GoogleProject(projectName)
+        implicit val token = ronAuthToken
+
+        val clusterName = ClusterName("user-jupyter-ext" + makeRandomId())
+        withNewCluster(project, clusterName, ClusterRequest(Map(), Option(testJupyterExtensionUri), None)) { cluster =>
+          withNewNotebook(cluster) { notebookPage =>
+            notebookPage.executeCell("1 + 1") shouldBe Some("2")
+            //Check if the mark up was translated correctly
+            notebookPage.translateMarkup("Hello") should include("Bonjour")
+          }
+        }
+      }
     }
 
   }
