+++ automation/src/test/scala/org/broadinstitute/dsde/workbench/leonardo/NotebookInteractionSpec.scala
@@ -287,20 +287,6 @@ class NotebookInteractionSpec extends FreeSpec with LeonardoTestUtils with Befor
       }
     }
 
-    //Test to check if extensions are installed correctly
-    //Using nbtranslate extension from here:
-    //https://github.com/ipython-contrib/jupyter_contrib_nbextensions/tree/master/src/jupyter_contrib_nbextensions/nbextensions/nbTranslate
-    "should install user specified notebook extensions" in withWebDriver { implicit driver =>
-      val clusterName = ClusterName("user-jupyter-ext" + makeRandomId())
-      withNewCluster(billingProject, clusterName, ClusterRequest(Map(), Option(testJupyterExtensionUri), None)) { cluster =>
-        withNewNotebook(cluster) { notebookPage =>
-          notebookPage.executeCell("1 + 1") shouldBe Some("2")
-          //Check if the mark up was translated correctly
-          notebookPage.translateMarkup("Hello") should include("Bonjour")
-        }
-      }
-    }
-
     Seq(Python2, Python3).foreach { kernel =>
       s"should be able to pip install packages using ${kernel.string}" in withWebDriver { implicit driver =>
         withNewNotebook(ronCluster, kernel) { notebookPage =>
